Refactor common to avoid having to publicly expose event_handlers with pub use
