+++ examples/common/basic/mod.rs
@@ -1,11 +1,9 @@
 use serde::{Deserialize, Serialize};
 
 use esrs::Aggregate;
-pub use event_handler::*;
-pub use view::*;
 
-mod event_handler;
-mod view;
+pub mod event_handler;
+pub mod view;
 
 #[derive(Clone)]
 pub struct BasicAggregate;
