+++ examples/common/shared/mod.rs
@@ -1,5 +1,2 @@
-pub use event_handler::*;
-pub use view::*;
-
-mod event_handler;
-mod view;
+pub mod event_handler;
+pub mod view;
