+++ examples/event_bus/main.rs
@@ -20,7 +20,10 @@ use esrs::manager::AggregateManager;
 use esrs::store::postgres::{PgStore, PgStoreBuilder, PgStoreError};
 use esrs::AggregateState;
 
-use crate::common::{new_pool, random_letters, BasicAggregate, BasicCommand, BasicError, BasicEventHandler, BasicView};
+use crate::common::basic::event_handler::BasicEventHandler;
+use crate::common::basic::view::BasicView;
+use crate::common::basic::{BasicAggregate, BasicCommand, BasicError};
+use crate::common::util::{new_pool, random_letters};
 use crate::kafka::KafkaEventBusConsumer;
 use crate::rabbit::RabbitEventBusConsumer;
