+++ examples/locking_strategies/main.rs
@@ -45,7 +45,8 @@ use esrs::store::postgres::{PgStore, PgStoreBuilder, PgStoreError};
 use esrs::store::EventStore;
 use esrs::AggregateState;
 
-use crate::common::{new_pool, BasicAggregate, BasicCommand, BasicError, BasicEvent};
+use crate::common::basic::{BasicAggregate, BasicCommand, BasicError, BasicEvent};
+use crate::common::util::new_pool;
 
 #[path = "../common/lib.rs"]
 mod common;
